+++ common/src/main/java/com/verdantartifice/primalmagick/common/items/wands/SpellScrollItem.java
@@ -58,7 +58,7 @@ public class SpellScrollItem extends Item {
     @Override
     public void appendHoverText(ItemStack stack, Item.TooltipContext context, List<Component> tooltip, TooltipFlag flag) {
         super.appendHoverText(stack, context, tooltip, flag);
-        tooltip.addAll(SpellManager.getSpellPackageDetailTooltip(this.getSpell(stack), stack, false, context.registries()));
+        tooltip.addAll(SpellManager.getSpellPackageDetailTooltip(this.getSpell(stack), stack, null, false, context.registries()));
         tooltip.add(TOOLTIP);
     }