+++ changelog.md
@@ -10,6 +10,7 @@ v6.0.6
 * Restored the Treasure enchantment to full functionality
 * Made several blocks waterloggable (Thanks, eehunter!)
 * Removed outdated references to mana cost penalties from some Grimoire entries
+* Fixed a tooltip bug for wands that were showing incorrect mana costs for spells
 
 v6.0.5
 ------